# docs/openapi-avatar-v1.yaml
openapi: 3.1.0
info:
  title: WordYeah Avatar Moderation API
  version: 0.1.0
  description: Local, fail-closed avatar moderation and review API.
servers:
  - url: http://127.0.0.1:18765
security:
  - bearerAuth: []
paths:
...
  /v1/moderate/image:
    post:
      summary: Moderate one local image payload.
      description: Raw image bytes only; no remote URL fetch.
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
          image/gif: { schema: { type: string, format: binary } }
          image/bmp: { schema: { type: string, format: binary } }
      responses:
        '200':
          description: Moderation result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ModerationResult' }
        '400': { description: Invalid request. }
        '401': { description: Missing or invalid API key. }
        '413': { description: Body exceeds configured limit. }
        '415': { description: Unsupported media type. }
        '422': { description: Model or media processing error; fail closed. }
        '503': { description: Model is not ready. }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    reviewerSession:
      type: apiKey
      in: cookie
      name: wordyeah_review_session
...
  schemas:
    Finding:
      type: object
      required: [category, label, source]
      properties:
        category: { type: string }
        label: { type: string }
        score: { type: [number, 'null'], minimum: 0, maximum: 1 }
        source: { type: string }
    ModerationResult:
      type: object
      required: [request_id, content_sha256, media_type, decision, reasons, findings]
      properties:
        request_id: { type: string }
        content_sha256: { type: string, pattern: '^[a-f0-9]{64}$' }
        media_type: { type: string, enum: [image] }
        decision: { type: string, enum: [allow, block, review, error] }
        reasons: { type: array, items: { type: string } }
        findings: { type: array, items: { $ref: '#/components/schemas/Finding' } }
        top_score: { type: [number, 'null'], minimum: 0, maximum: 1 }
        model_versions: { type: object, additionalProperties: { type: string } }
        elapsed_ms: { type: number, minimum: 0 }
        error: { type: [string, 'null'] }
...